Price import - Some form cleanup
