+++ src/Form/PriceImport/CustomerOptionValuePriceType.php
@@ -13,25 +13,18 @@ declare(strict_types=1);
 namespace Brille24\SyliusCustomerOptionsPlugin\Form\PriceImport;
 
 use Brille24\SyliusCustomerOptionsPlugin\Entity\CustomerOptions\CustomerOptionInterface;
-use Brille24\SyliusCustomerOptionsPlugin\Entity\CustomerOptions\CustomerOptionValueInterface;
 use Brille24\SyliusCustomerOptionsPlugin\Entity\CustomerOptions\CustomerOptionValuePrice;
-use Brille24\SyliusCustomerOptionsPlugin\Entity\Tools\DateRange;
 use Brille24\SyliusCustomerOptionsPlugin\Enumerations\CustomerOptionTypeEnum;
+use Brille24\SyliusCustomerOptionsPlugin\Form\Product\ProductCustomerOptionValuePriceTypeTrait;
 use Brille24\SyliusCustomerOptionsPlugin\Repository\CustomerOptionRepositoryInterface;
-use Sonata\CoreBundle\Form\Type\DateTimeRangeType;
-use Sylius\Bundle\ChannelBundle\Form\Type\ChannelChoiceType;
-use Sylius\Bundle\MoneyBundle\Form\Type\MoneyType;
-use Sylius\Component\Core\Model\ChannelInterface;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\CallbackTransformer;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
-use Symfony\Component\Form\Extension\Core\Type\PercentType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Intl\Currencies;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 final class CustomerOptionValuePriceType extends AbstractType
 {
+    use ProductCustomerOptionValuePriceTypeTrait;
+
     /** @var CustomerOptionRepositoryInterface */
     protected $customerOptionRepository;
 
@@ -40,6 +33,7 @@ final class CustomerOptionValuePriceType extends AbstractType
         $this->customerOptionRepository = $customerOptionRepository;
     }
 
+    /** {@inheritdoc} */
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $values = [];
@@ -50,89 +44,21 @@ final class CustomerOptionValuePriceType extends AbstractType
             }
         }
 
-        $builder
-            ->add('customerOptionValue', ChoiceType::class, [
-                'choices'      => $values,
-                'choice_label' => 'name',
-                'group_by'     => function (CustomerOptionValueInterface $customerOptionValue) {
-                    return $customerOptionValue->getCustomerOption()->getName();
-                },
-            ])
-            ->add('channel', ChannelChoiceType::class, [
-                'choice_attr' => static function (?ChannelInterface $channel) {
-                    if ($channel !== null) {
-                        if ($channel->getBaseCurrency() !== null) {
-                            $currency = $channel->getBaseCurrency()->getCode() ?? 'EUR';
-                        } else {
-                            $currency  = 'EUR';
-                        }
-                        $symbol = Currencies::getSymbol($currency, 'en');
-
-                        return ['data-attribute' => $symbol];
-                    }
-
-                    return '';
-                },
-                'attr' => ['onChange' => 'customerOptions.changeCustomerAmountCurrencyOnChannelChange(this);'],
-            ])
-            ->add('percent', PercentType::class, [
-                'empty_data' => '0.00',
-                'scale'      => 5,
-                'required'   => false,
-            ])
-            ->add('amount', MoneyType::class, [
-                'empty_data' => '0.00',
-                'currency'   => 'USD',
-                'required'   => false,
-            ])
-            ->add('type', ChoiceType::class, [
-                'choices'      => CustomerOptionValuePrice::getAllTypes(),
-                'choice_label' => static function ($option) {
-                    return 'brille24.ui.pricing.'.strtolower($option);
-                },
-            ])
-            ->add('dateValid', DateTimeRangeType::class, [
-                'required' => false,
-                'label'    => 'Active range',
-            ]);
-
-        $this->addModelTransformer($builder);
+        $this->addValuePriceFields($builder, $values);
     }
 
-    private function addModelTransformer(FormBuilderInterface $builder): void
-    {
-        $builder->get('dateValid')->addModelTransformer(
-            new CallbackTransformer(
-                static function (?DateRange $dateRange) {
-                    if ($dateRange === null) {
-                        return [];
-                    }
 
-                    return [
-                        'start' => $dateRange->getStart(),
-                        'end'   => $dateRange->getEnd(),
-                    ];
-                },
-                static function (array $dateTime) {
-                    if ($dateTime['start'] === null || $dateTime['end'] === null) {
-                        return null;
-                    }
-
-                    return new DateRange($dateTime['start'], $dateTime['end']);
-                }
-            )
-        );
-    }
-
+    /** {@inheritdoc} */
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver
             ->setDefaults([
-                  'data_class' => CustomerOptionValuePrice::class,
-              ])
+                'data_class' => CustomerOptionValuePrice::class,
+            ])
         ;
     }
 
+    /** {@inheritdoc} */
     public function getBlockPrefix(): string
     {
         return 'brille24_product_value_price';
